corrijo type ID a Long en repository:Docente,Preceptor,Tutor

+++ src/main/java/com/grup14/luterano/repository/DocenteRepository.java
@@ -3,5 +3,5 @@ package com.grup14.luterano.repository;
 import com.grup14.luterano.entities.Docente;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DocenteRepository extends JpaRepository<Docente,Integer> {
+public interface DocenteRepository extends JpaRepository<Docente,Long> {
 }
