+++ src/main/java/com/grup14/luterano/repository/PreceptorRepository.java
@@ -3,5 +3,5 @@ package com.grup14.luterano.repository;
 import com.grup14.luterano.entities.Preceptor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PreceptorRepository extends JpaRepository<Preceptor,Integer> {
+public interface PreceptorRepository extends JpaRepository<Preceptor,Long> {
 }
