+++ src/main/java/com/grup14/luterano/repository/TutorRepository.java
@@ -3,5 +3,5 @@ package com.grup14.luterano.repository;
 import com.grup14.luterano.entities.Tutor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TutorRepository extends JpaRepository<Tutor,Integer> {
+public interface TutorRepository extends JpaRepository<Tutor,Long> {
 }
